Refactor: Extract DomTester
